moved all report related route to a single blueprint
